requirements.txt
flask==2.3.2
flask_cors==4.0.0
werkzeug==2.3.6
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
opencv-python==4.7.0.72
Pillow==10.0.0
tqdm==4.65.0
scikit-learn==1.2.2
timm==0.9.2
gunicorn==20.1.0
boto3==1.28.38
python-dotenv==1.0.0
